MongoDB update!
Now we are connected to a database in MongoDB so the daily restart of Heroku servers will no longer affect the bot. The IDs of the answered tweets will be stored on this database.
In the past, the daily Heroku servers restart generated a bug in the bot, responding to a tweet already answered. Now it won´t happen again! :D